v0.32 WIP
Work in progress build switching to gsKit, Not fully functional
