+++ VTSPS2-HBDL.h
@@ -1,9 +1,12 @@
+#include <gsKit.h>
+#include <dmaKit.h>
+#include <malloc.h>
+#include <gsToolkit.h>
 #include <tamtypes.h>
 #include <errno.h>
 #include <kernel.h>
 #include <sifrpc.h>
 #include <loadfile.h>
-//#include <fileio.h>
 #include <libmc.h>
 #include <stdio.h>
 #include <string.h>
@@ -15,12 +18,9 @@
 #include "malloc.h"
 #include <libcdvd.h>
 #include <iopheap.h>
-//#include <io_common.h>
 #include <fcntl.h>
 #include <syscallnr.h>
 #include <unistd.h>
-//#include <fileXio.h>
-//#include <fileXio_rpc.h>
 #include "checksum.h"
 
 // IP Shit.
@@ -30,9 +30,21 @@
 // SBV
 #include <sbv_patches.h>
 
-
 //ps2ip.c
 int dhcpmain();
+
+//VTSPS2-CRC32.c
+char* file_crc32(char device[], char path[], char fn[]);
+
+//misc.c
+void substring(char s[], char sub[], int p, int l);
+int getFileSize(int fd);
+void LoadElf(const char *elf, char* path);
+ssize_t getline(char **lineptr, size_t *n, FILE *stream);
+
+//pad.c
+void buttonStatts(int port, int slot);
+void checkPadConnected(void);
 
 // Modules
 extern void freesio2;
@@ -70,7 +82,7 @@ extern u32 size_usbhdfsd;
 
 //PAD VARIABLES
 //check for multiple definitions
-#define DEBUG
+//#define DEBUG
 
 #if !defined(ROM_PADMAN) && !defined(NEW_PADMAN)
 #define ROM_PADMAN
@@ -104,4 +116,3 @@ int LoadIRX();
 
 
 #define TYPE_XMC
-
