+++ misc.c
@@ -1,0 +1,431 @@
+#include "VTSPS2-HBDL.h"
+
+char *appFail = "Application Failure!\n";
+char *modloadfail = "Failed to load module: ";
+char *osdmsg = "Exiting to OSDSYS\n";
+
+int getFileSize(int fd)
+{
+	int size = lseek(fd, 0, SEEK_END);
+	lseek(fd, 0, SEEK_SET);
+	return size;
+}
+
+void ResetIOP()
+{
+
+	SifInitRpc(0);           //Initialize SIFRPC and SIFCMD. Although seemingly unimportant, this will update the addresses on the EE, which can prevent a crash from happening around the IOP reboot.
+	SifIopReset("", 0);      //Reboot IOP with default modules (empty command line)
+	while(!SifIopSync()){}   //Wait for IOP to finish rebooting.
+	SifInitRpc(0);           //Initialize SIFRPC and SIFCMD.
+	SifLoadFileInit();       //Initialize LOADFILE RPC.
+	// SBV Patches Are Not part of a Normal IOP Reset.
+	sbv_patch_enable_lmb(); //SBV Patches
+	sbv_patch_disable_prefix_check(); //SBV Patch Load Executable IRX And ELF Files From User-Writable Storage
+	//sbv_patch_user_mem_clear(0x00100000); // You Can Specify a Starting Address for the Wipe
+	//sbv_patch_user_mem_clear(0x02000000); // Disable Clear Memory With LoadExecPS2() when 0x02000000 is passed as an arg
+}
+
+void gotoOSDSYS(int sc)
+{
+	if (sc != 0)
+	{
+		scr_printf(appFail);
+		if(sc ==1 || sc ==2 || sc ==3 || sc ==4 || sc ==5)
+		{
+			scr_printf(modloadfail);
+		}
+		if (sc == 1)
+		{
+			scr_printf("SIO2MAN\n");
+		}
+		if (sc == 2)
+		{
+			scr_printf("CDVDMAN\n");
+		}
+		if (sc == 3)
+		{
+			scr_printf("PADMAN\n");
+		}
+		if (sc == 4)
+		{
+			scr_printf("MCMAN\n");
+		}
+		if (sc == 5)
+		{
+			scr_printf("MCSERV\n");
+		}
+		if (sc == 6)
+		{
+			scr_printf("ERROR: Unknown\n");
+		}
+		//sleep(5);
+	}
+	ResetIOP();
+	scr_printf(osdmsg);
+	LoadExecPS2("rom0:OSDSYS", 0, NULL);
+}
+
+void LoadModules(void)
+{
+	int ret;
+
+	ret = SifExecModuleBuffer(&freesio2, size_freesio2, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf("Failed to Load freesio2 sw module");
+		gotoOSDSYS(1);
+	}
+
+	ret = SifExecModuleBuffer(&iomanX, size_iomanX, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf("Failed to Load iomanx sw module");
+	}
+
+	//ret = SifExecModuleBuffer(&fileXio, size_fileXio, 0, NULL, NULL);
+	//if (ret < 0)
+	//{
+	//	scr_printf("Failed to Load freesio2 sw module");
+	//}
+
+
+	ret = SifExecModuleBuffer(&freepad, size_freepad, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf("Failed to Load freepad sw module");
+		gotoOSDSYS(3);
+	}
+
+	ret = SifExecModuleBuffer(&mcman, size_mcman, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf("Failed to Load mcman sw module");
+		gotoOSDSYS(4);
+	}
+
+	ret = SifExecModuleBuffer(&mcserv, size_mcserv, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf("Failed to Load mcserv sw module");
+		gotoOSDSYS(5);
+
+	}
+
+	ret = SifExecModuleBuffer(&ps2dev9, size_ps2dev9, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf(" Could not load ps2dev9.IRX! %d\n", ret);
+		SleepThread();
+	}
+
+
+	ret = SifExecModuleBuffer(&netman, size_netman, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf(" Could not load netman.IRX! %d\n", ret);
+		SleepThread();
+	}
+
+	ret = SifExecModuleBuffer(&smap, size_smap, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf(" Could not load smap.IRX! %d\n", ret);
+		SleepThread();
+	}
+
+	ret = SifExecModuleBuffer(&ps2ipnm, size_ps2ipnm, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf(" Could not load ps2ip.IRX! %d\n", ret);
+		SleepThread();
+	}
+
+	ret = SifExecModuleBuffer(&ps2ips, size_ps2ips, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf(" Could not load ps2ips.IRX! %d\n", ret);
+		SleepThread();
+	}
+	ps2ip_init();
+
+	ret = SifExecModuleBuffer(&ps2http, size_ps2http, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf(" Could not load ps2http.IRX! %d\n", ret);
+		gotoOSDSYS(5);
+	}
+
+	ret = SifExecModuleBuffer(&usbd, size_usbd, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf(" Could not load usbd.irx! %d\n", ret);
+		SleepThread();
+	}
+
+	ret = SifExecModuleBuffer(&usbhdfsd, size_usbhdfsd, 0, NULL, NULL);
+	if (ret < 0)
+	{
+		scr_printf(" Could not load usbhdfsd.irx! %d\n", ret);
+		SleepThread();
+	}
+}
+
+//Function LoadElf() from main.c MPLUS-LOADER3.ELF
+//http://lukasz.dk/2008/04/22/datel-memory-plus-64-mb/
+//slightly modified
+void LoadElf(const char *elf, char* path)
+{
+	char* args[1];
+	t_ExecData exec;
+
+	SifLoadElf(elf, &exec);
+
+	#ifdef DEBUG
+		//scr_printf("Trying to load ELF: %s\n", elf);
+	#endif
+
+	if(exec.epc > 0)
+	{
+		FlushCache(0);
+		FlushCache(2);
+
+		// Reset IOP, since not all ELF's do it and we've loaded additional IOP
+		// modules which need to be unloaded
+		ResetIOP();
+
+		if(path != 0)
+		{
+			args[0] = path;
+			ExecPS2((void*)exec.epc, (void*)exec.gp, 1, args);
+		}
+		else
+		{
+			ExecPS2((void*)exec.epc, (void*)exec.gp, 0, NULL);
+		}
+	}
+}
+
+////////////////
+//initalizePad//
+////////////////
+
+static int initializePad(int port, int slot)
+{
+
+	int ret;
+	int modes;
+	int i;
+
+	waitPadReady(port, slot);
+	modes = padInfoMode(port, slot, PAD_MODETABLE, -1);
+	if (modes > 0) {
+		for (i = 0; i < modes; i++) {
+		}
+
+	}
+	if (modes == 0) {
+		return 1;
+	}
+
+	i = 0;
+	do {
+		if (padInfoMode(port, slot, PAD_MODETABLE, i) == PAD_TYPE_DUALSHOCK)
+			break;
+		i++;
+	} while (i < modes);
+	if (i >= modes) {
+		return 1;
+	}
+
+	ret = padInfoMode(port, slot, PAD_MODECUREXID, 0);
+	if (ret == 0) {
+		return 1;
+	}
+	padSetMainMode(port, slot, PAD_MMODE_DUALSHOCK, PAD_MMODE_LOCK);
+
+	waitPadReady(port, slot);
+	padInfoPressMode(port, slot);
+
+	waitPadReady(port, slot);
+	padEnterPressMode(port, slot);
+
+	waitPadReady(port, slot);
+	actuators = padInfoAct(port, slot, -1, 0);
+
+	if (actuators != 0) {
+		actAlign[0] = 0;
+		actAlign[1] = 1;
+		actAlign[2] = 0xff;
+		actAlign[3] = 0xff;
+		actAlign[4] = 0xff;
+		actAlign[5] = 0xff;
+
+		waitPadReady(port, slot);
+
+		padSetActAlign(port, slot, actAlign);
+	}
+	else {
+	}
+	return 1;
+}
+
+////////////////
+//buttonStatts//
+////////////////
+void buttonStatts(int port, int slot)
+{
+	int ret;
+	ret = padRead(port, slot, &buttons);
+
+	if (ret != 0) {
+		paddata = 0xffff ^ buttons.btns;
+
+		new_pad = paddata & ~old_pad;
+		old_pad = paddata;
+	}
+}
+
+/////////////////////
+//checkPadConnected//
+/////////////////////
+void checkPadConnected(void)
+{
+	int ret, i;
+	ret = padGetState(0, 0);
+	while ((ret != PAD_STATE_STABLE) && (ret != PAD_STATE_FINDCTP1)) {
+		if (ret == PAD_STATE_DISCONN) {
+			#if defined DEBUG
+				scr_printf("Controller(%d, %d) is disconnected\n", 0, 0);
+			#endif
+		}
+		ret = padGetState(0, 0);
+	}
+	if (i == 1) {
+	}
+}
+
+///////////////////
+//pad_wait_button//
+///////////////////
+void pad_wait_button(u32 button)
+{
+	while (1)
+	{
+		buttonStatts(0, 0);
+		if (new_pad & button) return;
+	}
+}
+
+////////////////
+//waitPadReady//
+////////////////
+int waitPadReady(int port, int slot)
+{
+	int state;
+	int lastState;
+	char stateString[16];
+
+	state = padGetState(port, slot);
+	lastState = -1;
+	while ((state != PAD_STATE_STABLE) && (state != PAD_STATE_FINDCTP1)) {
+		if (state != lastState) {
+			padStateInt2String(state, stateString);
+		}
+		lastState = state;
+		state = padGetState(port, slot);
+	}
+	// Were the pad ever 'out of sync'?
+	if (lastState != -1) {
+
+	}
+	return 0;
+}
+
+void initialize(void)
+{
+
+	int ret;
+	// Initialize
+	SifInitRpc(0);
+	ResetIOP();	
+	//scr_clear();
+	// init debug screen
+	init_scr();
+	scr_clear();
+	scr_printf("Loading... Please Wait. \n");
+	// load all modules
+	LoadModules();
+	// init pad
+	padInit(0);
+	if ((ret = padPortOpen(0, 0, padBuf)) == 0)
+	{
+		#if defined DEBUG
+			scr_printf("padOpenPort failed: %d\n", ret);
+		#endif
+		SleepThread();
+	}
+	if (!initializePad(0, 0))
+	{
+		#if defined DEBUG
+			scr_printf("pad initalization failed!\n");
+		#endif
+		SleepThread();
+
+	}
+}
+
+// This function is public domain -- Will Hartung 4/9/09 */
+// Modifications, public domain as well, by Antti Haapala, 11/10/17
+// - Switched to getc on 5/23/19 */
+
+// if typedef doesn't exist (msvc, blah)
+//typedef intptr_t ssize_t;
+ssize_t getline(char **lineptr, size_t *n, FILE *stream)
+{
+    size_t pos;
+    int c;
+
+    if (lineptr == NULL || stream == NULL || n == NULL) {
+        errno = EINVAL;
+        return -1;
+    }
+
+    c = getc(stream);
+    if (c == EOF) {
+        return -1;
+    }
+
+    if (*lineptr == NULL) {
+        *lineptr = malloc(128);
+        if (*lineptr == NULL) {
+            return -1;
+        }
+        *n = 128;
+    }
+
+    pos = 0;
+    while(c != EOF) {
+        if (pos + 1 >= *n) {
+            size_t new_size = *n + (*n >> 2);
+            if (new_size < 128) {
+                new_size = 128;
+            }
+            char *new_ptr = realloc(*lineptr, new_size);
+            if (new_ptr == NULL) {
+                return -1;
+            }
+            *n = new_size;
+            *lineptr = new_ptr;
+        }
+
+        ((unsigned char *)(*lineptr))[pos ++] = c;
+        if (c == '\n') {
+            break;
+        }
+        c = getc(stream);
+    }
+
+    (*lineptr)[pos] = '\0';
+    return pos;
+}
