+++ pad.c
@@ -1,0 +1,137 @@
+#include "VTSPS2-HBDL.h"
+
+////////////////
+//initalizePad//
+////////////////
+
+static int initializePad(int port, int slot)
+{
+
+	int ret;
+	int modes;
+	int i;
+
+	waitPadReady(port, slot);
+	modes = padInfoMode(port, slot, PAD_MODETABLE, -1);
+	if (modes > 0) {
+		for (i = 0; i < modes; i++) {
+		}
+
+	}
+	if (modes == 0) {
+		return 1;
+	}
+
+	i = 0;
+	do {
+		if (padInfoMode(port, slot, PAD_MODETABLE, i) == PAD_TYPE_DUALSHOCK)
+			break;
+		i++;
+	} while (i < modes);
+	if (i >= modes) {
+		return 1;
+	}
+
+	ret = padInfoMode(port, slot, PAD_MODECUREXID, 0);
+	if (ret == 0) {
+		return 1;
+	}
+	padSetMainMode(port, slot, PAD_MMODE_DUALSHOCK, PAD_MMODE_LOCK);
+
+	waitPadReady(port, slot);
+	padInfoPressMode(port, slot);
+
+	waitPadReady(port, slot);
+	padEnterPressMode(port, slot);
+
+	waitPadReady(port, slot);
+	actuators = padInfoAct(port, slot, -1, 0);
+
+	if (actuators != 0) {
+		actAlign[0] = 0;
+		actAlign[1] = 1;
+		actAlign[2] = 0xff;
+		actAlign[3] = 0xff;
+		actAlign[4] = 0xff;
+		actAlign[5] = 0xff;
+
+		waitPadReady(port, slot);
+
+		padSetActAlign(port, slot, actAlign);
+	}
+	else {
+	}
+	return 1;
+}
+
+////////////////
+//buttonStatts//
+////////////////
+static void buttonStatts(int port, int slot)
+{
+	int ret;
+	ret = padRead(port, slot, &buttons);
+
+	if (ret != 0) {
+		paddata = 0xffff ^ buttons.btns;
+
+		new_pad = paddata & ~old_pad;
+		old_pad = paddata;
+	}
+}
+
+/////////////////////
+//checkPadConnected//
+/////////////////////
+void checkPadConnected(void)
+{
+	int ret, i;
+	ret = padGetState(0, 0);
+	while ((ret != PAD_STATE_STABLE) && (ret != PAD_STATE_FINDCTP1)) {
+		if (ret == PAD_STATE_DISCONN) {
+			#if defined DEBUG
+				scr_printf("Controller(%d, %d) is disconnected\n", 0, 0);
+			#endif
+		}
+		ret = padGetState(0, 0);
+	}
+	if (i == 1) {
+	}
+}
+
+///////////////////
+//pad_wait_button//
+///////////////////
+void pad_wait_button(u32 button)
+{
+	while (1)
+	{
+		buttonStatts(0, 0);
+		if (new_pad & button) return;
+	}
+}
+
+////////////////
+//waitPadReady//
+////////////////
+static int waitPadReady(int port, int slot)
+{
+	int state;
+	int lastState;
+	char stateString[16];
+
+	state = padGetState(port, slot);
+	lastState = -1;
+	while ((state != PAD_STATE_STABLE) && (state != PAD_STATE_FINDCTP1)) {
+		if (state != lastState) {
+			padStateInt2String(state, stateString);
+		}
+		lastState = state;
+		state = padGetState(port, slot);
+	}
+	// Were the pad ever 'out of sync'?
+	if (lastState != -1) {
+
+	}
+	return 0;
+}
